--- next-app/src/__server-express__/controller/index.ts
@@ -1,2 +1,0 @@
-export { register, login, logout, session, me } from "./auth.controller.js"
-export { createProject, getProjects, updateProject, deleteProject } from "./project.controller.js"
