--- next-app/src/__server-express__/lib/generateToken.ts
@@ -1,6 +1,0 @@
-import { signToken } from "./jwt.js";
-import { JWTPayload } from "../types/types.payload.js";
-
-export const generateToken = (data: JWTPayload): string => {
-  return signToken({ email: data.email });
-}
