+++ next-app/src/server/error/AuthError.ts
