+++ next-app/src/server/error/SecurityError.ts
