+++ next-app/src/server/lib/bcryptPassword.ts
@@ -1,10 +1,10 @@
-import bcrypt from "bcrypt"
+import bcrypt from "bcrypt";
 import { InvalidPasswordError } from "../error/index.js";
-import { RegisterRequest } from "@pkg/shared";
+import { DbUser } from "../types";
 
 const SaltRounds = 10;
 
-export const hashPassword = async (password: RegisterRequest["password"]) => {
+export const hashPassword = async (password: DbUser["passwordHash"]) => {
   return await bcrypt.hash(password, SaltRounds);
 }
