+++ next-app/src/server/lib/index.ts
@@ -1,3 +1,2 @@
-export { signToken, verifyToken } from "./jwt.js"
 export { hashPassword, comparePassword } from "./bcryptPassword.js"
 export { prisma } from "./prisma.js"
