+++ next-app/tsconfig.json
@@ -20,7 +20,8 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@/schemas": ["./src/schemas/index.ts"]
+      "@/schemas": ["./src/schemas/index.ts"],
+      "@/server": ["./src/server/index.ts"]
     }
   },
   "include": [
